Automatically backed up by Learn

+++ db/migrate/001_create_users_table.rb
@@ -3,7 +3,7 @@ class CreateUsersTable < ActiveRecord::Migration[5.1]
     create_table :users do |x|
       x.string :username
       x.string :password 
-      x.integer :balance
+      x.decimal :balance
     end
   end
 end
